+++ src/collective/eeafaceted/z3ctable/columns.py
@@ -409,7 +409,7 @@ class AbbrColumn(VocabularyColumn):
 
     def renderCell(self, item):
         value = self.getValue(item)
-        if not value:
+        if not value or value == self.ignored_value:
             return u'-'
 
         # caching when several same values in same column